requirements.txt
numpy==2.0.0
bosdyn-client==4.1.1
bosdyn-mission==4.1.1
bosdyn-choreography-client==4.1.1
bosdyn-orbit==4.1.1
opencv-python
scipy==1.13.1
